Added implementation for C++ version of ncurses, revised base classes to conform

+++ Ex1/CircularBoard.cpp
@@ -9,7 +9,7 @@
 #include "CircularBoard.hpp"
 #include "Tile.hpp"
 
-CircularBoard::CircularBoard(size_t width, size_t height) :
+CircularBoard::CircularBoard(int width, int height) :
 Board(width, height)
 { }
